+++ src/components/Reservations/Card/Card.tsx
@@ -5,39 +5,10 @@ import formatDateString from '@/utils/formatDateString';
 import CancelModal from '@/components/Modal/CancelModal/CancelModal';
 import { useState } from 'react';
 import ReviewModal from '@/components/Modal/ReviewModal/ReviewModal';
+import { Reservations } from '@/types/reservations';
+import { STATUS } from '@/constants/reservation';
 
-type Activity = {
-  bannerImageUrl: string;
-  title: string;
-  id: number;
-};
-
-const STATUS = {
-  pending: '예약 완료',
-  confirmed: '예약 승인',
-  declined: '예약 거절',
-  canceled: '예약 취소',
-  completed: '체험 완료',
-};
-
-type Status = keyof typeof STATUS;
-
-export interface CardProps {
-  data: {
-    id: number;
-    activity: Activity;
-    scheduleId: number;
-    status: Status;
-    reviewSubmitted: boolean;
-    totalPrice: number;
-    headCount: number;
-    date: string;
-    startTime: string;
-    endTime: string;
-  };
-}
-
-function Card({ data }: CardProps) {
+function Card({ data }: Reservations) {
   const [isCancelModalOpen, setIsCancelModalOpen] = useState(false);
   const [isReviewModalOpen, setIsReviewModalOpen] = useState(false);
   const handleCancelModalToggle = () => {
@@ -73,19 +44,19 @@ function Card({ data }: CardProps) {
               예약 취소
             </button>
           )}
-          {/* TODO : api 연결 후 handleCancel에 적절한 함수 연결 필요 */}
-          {isCancelModalOpen && (
-            <CancelModal handleModalClose={handleCancelModalToggle} handleCancel={handleCancelModalToggle} />
-          )}
           {/* 체험 완료일 때만 후기 작성 버튼 보이고, reviewSubmit이 true면 disabled */}
           {data.status === 'completed' && (
             <button onClick={handleReviewModalToggle} disabled={data.reviewSubmitted} className={styles.button}>
               후기 작성
             </button>
           )}
-          {isReviewModalOpen && <ReviewModal handleModalClose={handleReviewModalToggle} data={data} />}
         </div>
       </div>
+      {/* TODO : api 연결 후 handleCancel에 적절한 함수 연결 필요 */}
+      {isCancelModalOpen && (
+        <CancelModal handleModalClose={handleCancelModalToggle} handleCancel={handleCancelModalToggle} />
+      )}
+      {isReviewModalOpen && <ReviewModal handleModalClose={handleReviewModalToggle} data={data} />}
     </div>
   );
 }
