+++ src/constants/reservation.ts
@@ -1,0 +1,7 @@
+export const STATUS = {
+  pending: '예약 완료',
+  confirmed: '예약 승인',
+  declined: '예약 거절',
+  canceled: '예약 취소',
+  completed: '체험 완료',
+};
